% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{shiny_push}
\alias{shiny_push}
\title{将本地的程序发布到reshapedata.com上}
\usage{
shiny_push(appName = "", parent = "")
}
\arguments{
\item{appName}{程序目录名称}

\item{parent}{上述路径名称}
}
\value{
返回值
}
\description{
将本地的程序发布到reshapedata.com上
}
\examples{
shiny_push('demo1','demo');
}
